+++ app/routes/post.py
@@ -1,0 +1,98 @@
+"""
+app.routes.post
+===============
+"""
+from datetime import datetime
+from typing import Optional, Union
+
+from flask import Blueprint, render_template
+from flask_login import current_user, login_required
+from werkzeug import Response
+
+from app import redirect
+from app.forms import EmptyForm, PostForm
+from app.models import Post, db
+from app.security import authorization_required
+
+blueprint = Blueprint("post", __name__, url_prefix="/post")
+
+
+@blueprint.route("/create", methods=["GET", "POST"])
+@login_required
+@authorization_required
+def create() -> Union[str, Response]:
+    """Create a post.
+
+    The decorator will ensure that the user is logged in to visit this
+    view, otherwise they will be redirected the login page.
+
+    The create view works the same as the auth register view. The posted
+    data is validated and either the form is displayed (and the post is
+    added to the database) or an error is shown.
+
+    :return: Rendered create template on GET or failed POST. Response
+        object redirect to index view on successful POST.
+    """
+    form = PostForm()
+    if form.validate_on_submit():
+        post = Post(
+            title=form.title.data, body=form.body.data, user_id=current_user.id
+        )
+        db.session.add(post)
+        db.session.commit()
+        return redirect.index()
+
+    return render_template("post/create.html", form=form)
+
+
+@blueprint.route("/post/<int:id>", methods=["GET"])
+def post_page(id: int) -> str:
+    """Render post page for selected post ID.
+
+    :param id: ID of post to display full page on.
+    :return: Rendered post template.
+    """
+    post = Post.get_post(id, checkauthor=False)
+    return render_template("post/post.html", post=post)
+
+
+@blueprint.route("/<int:id>/update/", methods=["GET", "POST"])
+@blueprint.route("/<int:id>/update/<int:revision>", methods=["GET", "POST"])
+@login_required
+@authorization_required
+def update(id: int, revision: Optional[int] = None) -> Union[str, Response]:
+    """Update post that corresponds to the provided post ID.
+
+    :param id: The post's ID.
+    :param revision: Version to revert to.
+    :return: Rendered update template on GET or failed POST. Response
+        object redirect to index view on successful update POST.
+    """
+    post = Post.get_post(id, revision)
+    form = PostForm(title=post.title, body=post.body)
+    if form.validate_on_submit():
+        post.title = form.title.data
+        post.body = form.body.data
+        post.edited = datetime.utcnow()
+        db.session.commit()
+        return redirect.index()
+
+    return render_template("post/update.html", post=post, form=form)
+
+
+@blueprint.route("/<int:id>/version/<int:revision>")
+@login_required
+@authorization_required
+def version(id: int, revision: int) -> Union[str, Response]:
+    """Rewind versioned post that corresponds to the provided post ID.
+
+    :param id: The post's ID.
+    :param revision: Version to revert to.
+    :return: Rendered update template on GET or failed POST. Response
+        object redirect to index view on successful update POST.
+    """
+    post = Post.get_post(id, revision)
+    form = EmptyForm()
+    return render_template(
+        "post/post.html", post=post, id=id, revision=revision, form=form
+    )